+++ crates/kornia-apriltag/Cargo.toml
@@ -20,6 +20,7 @@ harness = false
 name = "bench_tagfamily"
 
 [dependencies]
+kornia-3d = { workspace = true }
 kornia-algebra = { workspace = true }
 kornia-image = { workspace = true }
 kornia-imgproc = { workspace = true }
